Fix dangerous casts
On Windows 64-bit, the size of long is 32-bit, thus any pointer to long cast is
not valid. In any case, one should use intptr_t and ptrdiff_t when casting
to integers. This commit attempts to fix all instances reported by GCC.
When relevant, I replaced code by the macros PTR_ADD, ALIGN_UP from system.h

+++ apps/misc.c
@@ -215,7 +215,7 @@ int fast_readline(int fd, char *buf, int buf_size, void *parameters,
         count++;
         if (next)
         {
-            pos = buf_size - ((long)next - (long)buf) - 1;
+            pos = buf_size - ((intptr_t)next - (intptr_t)buf) - 1;
             memmove(buf, next, pos);
         }
         else
